fix test

+++ runtime/runtime-params-estimator/src/function_call.rs
@@ -199,7 +199,7 @@ fn test_function_call_all_codes(metric: GasMetric, vm_kind: VMKind) {
                     as f64)
                     / 10f64.pow(12)
             } // teragas
-            GasMetric::Time => raw_result / (REPEATS * 1_000_000 as f64), // ms
+            GasMetric::Time => raw_result as f64 / (REPEATS as f64 * 1_000_000 as f64), // ms
         };
         estimations.push(Estimation {
             id: estimated_code.id.clone(),
